The slider is sliding!
Added a new texture for the label for wireframe line width.

+++ Innlevering2/GUI_Util.cpp
@@ -76,7 +76,7 @@ namespace gui
 		height = 0;
 	}
 
-	Rect::Rect( int x, int y, int width, int height )
+	Rect::Rect( float x, float y, float width, float height )
 	{
 		this->x = x;
 		this->y = y;
